changes in the add_one function inside the MySQLPipeline class
the function was concatenating strings with variables values in order to
form the SQL query. This will lead to serious problems with tweets
containing single or double quotes.

The function is now using prepared statement for better performance
while communicating with the MySQL Server and to avoid problems with
single and double quotes